TUNIC: Fix belltowers and laurels autosplits
This patch replaces the expected event IDs for the two belltowers and
for the Hero's Laurels to instead use the achievement event IDs. Due to
the way the autosplitter works and the way in which the last event ID is
exposed to the autosplitter by the game, we only see the most recent
event even if there were additional events between the previous update
and now. Since these three events also trigger achievement events, we
need to check for those in order for the autosplitter to work for them.