TUNIC: Fix sword split and add all relics split
This patch replaces the expected event ID for the sword with the
achievement event ID since the last event reported by the game is the
latter and we never see the sword event in the speedrun data.

A new autosplit is added for having restored corporeal form (as
described by the Steam achievement). This is required as the final relic
obtained will also trigger the achievement event before we get a chance
to see it in the last event field of the speedrun data.

Finally, the description of the hyperdash autosplit is updated for
clarity as the game refers to the item as the Hero's Laurels.